display fetched api data in HomeView

+++ CryptoTracker/Models/Coin.swift
@@ -69,7 +69,7 @@ struct Coin: Identifiable, Codable {
     let priceChange24H, priceChangePercentage24H: Double?
     let marketCapChange24H: Double?
     let marketCapChangePercentage24H: Double?
-    let circulatingSupply, totalSupply, maxSupply, ath: Int?
+    let circulatingSupply, totalSupply, maxSupply, ath: Double?
     let athChangePercentage: Double?
     let athDate: String?
     let atl, atlChangePercentage: Double?
